perf(solver): cache infer-match expansions
Cache pure evaluate_for_infer_match expansions inside the owning TypeEvaluator request. The key is the source/pattern TypeId under the evaluator's resolver and noUncheckedIndexedAccess mode; entries are dropped with the request and skipped when the fresh sub-evaluator reports a depth or silent-recursion bailout, so infer bindings and relation outcomes remain uncached.

AgentName: Studio-B
